feat: 🎸 Progress in service contribution

+++ src/API/MFFVP.Api/Application/Products/IPortfoliosService.cs
@@ -10,7 +10,7 @@ namespace MFFVP.Api.Application.Products
     public interface IPortfoliosService
     {
         Task<Result<IReadOnlyCollection<PortfolioResponse>>> GetPortfoliosAsync(ISender sender);
-        Task<Result<PortfolioResponse>> GetPortfolioAsync(long portfolioId, ISender sender);
+        Task<Result<PortfolioResponse>> GetPortfolioAsync(int portfolioId, ISender sender);
         Task<Result> CreatePortfolioAsync(CreatePortfolioCommand request, ISender sender);
         Task<Result> UpdatePortfolioAsync(UpdatePortfolioCommand request, ISender sender);
         Task<Result> DeletePortfolioAsync(long portfolioId, ISender sender);
