+++ src/Modules/Products/Domain/Alternatives/AlternativeCreatedDomainEvent.cs
@@ -2,7 +2,7 @@ using Common.SharedKernel.Domain;
 
 namespace Products.Domain.Alternatives;
 
-public sealed class AlternativeCreatedDomainEvent(long alternativeId) : DomainEvent
+public sealed class AlternativeCreatedDomainEvent(int alternativeId) : DomainEvent
 {
-    public long AlternativeId { get; } = alternativeId;
+    public int AlternativeId { get; } = alternativeId;
 }
