+++ src/Modules/Products/Domain/Alternatives/AlternativeErrors.cs
@@ -4,7 +4,7 @@ namespace Products.Domain.Alternatives;
 
 public static class AlternativeErrors
 {
-    public static Error NotFound(long alternativeId)
+    public static Error NotFound(int alternativeId)
     {
         return Error.NotFound(
             "Alternative.NotFound",
