+++ src/Modules/Products/Domain/Objectives/ObjectiveCreatedDomainEvent.cs
@@ -2,7 +2,7 @@ using Common.SharedKernel.Domain;
 
 namespace Products.Domain.Objectives;
 
-public sealed class ObjectiveCreatedDomainEvent(long objectiveId) : DomainEvent
+public sealed class ObjectiveCreatedDomainEvent(int objectiveId) : DomainEvent
 {
-    public long ObjectiveId { get; } = objectiveId;
+    public int ObjectiveId { get; } = objectiveId;
 }
