+++ src/Modules/Products/Domain/Portfolios/PortfolioErrors.cs
@@ -4,7 +4,7 @@ namespace Products.Domain.Portfolios;
 
 public static class PortfolioErrors
 {
-    public static Error NotFound(long portfolioId)
+    public static Error NotFound(int portfolioId)
     {
         return Error.NotFound(
             "Portfolio.NotFound",
